fase-5: desacoplamento da logica de validaçao para melhor testabilidade
