+++ src/main/java/br/com/fiaplanchespayment/domain/PaymentValidatorPort.java
@@ -1,0 +1,7 @@
+package br.com.fiaplanchespayment.domain;
+
+import br.com.fiaplanchespayment.application.dtos.PaymentOrderDto;
+
+public interface PaymentValidatorPort {
+    boolean validatePaymentOrder(PaymentOrderDto paymentOrderDto) throws Exception;
+}
